+++ tests/test_providers/test_internet.py
@@ -49,13 +49,15 @@ class TestInternet(object):
         result = net.image_placeholder(width=400, height=300)
         assert result is not None
 
-    def test_stock_image(self, net):
-        result = net.stock_image()
-        assert result is not None
-        result_2 = net.stock_image(category='nature').split('/')[-1][1:]
-        assert result_2 == 'nature'
-        result_3 = net.stock_image(width=1900, height=1080).split('/')[-2]
-        assert result_3 == '1900x1080'
+    @pytest.mark.parametrize(
+        'width, height, wh', [
+            (1920, 1080, '1920x1080'),
+            (800, 800, '800x800'),
+        ],
+    )
+    def test_stock_image(self, net, width, height, wh):
+        result = net.stock_image(width=width, height=height)
+        assert wh in result
 
     def test_image_by_keyword(self, net):
         result = net.image_by_keyword(keyword='word').split('/')[-1]
@@ -184,12 +186,12 @@ class TestSeededInternet(object):
     def test_home_page(self, i1, i2):
         assert i1.home_page() == i2.home_page()
         assert i1.home_page(tld_type=TLDType.GEOTLD) == \
-            i2.home_page(tld_type=TLDType.GEOTLD)
+               i2.home_page(tld_type=TLDType.GEOTLD)
 
     def test_subreddit(self, i1, i2):
         assert i1.subreddit() == i2.subreddit()
         assert i1.subreddit(nsfw=True, full_url=True) == \
-            i2.subreddit(nsfw=True, full_url=True)
+               i2.subreddit(nsfw=True, full_url=True)
 
     def test_user_agent(self, i1, i2):
         assert i1.user_agent() == i2.user_agent()
@@ -197,22 +199,20 @@ class TestSeededInternet(object):
     def test_image_placeholder(self, i1, i2):
         assert i1.image_placeholder() == i2.image_placeholder()
         assert i1.image_placeholder(width=128, height=128) == \
-            i2.image_placeholder(width=128, height=128)
+               i2.image_placeholder(width=128, height=128)
 
     def test_stock_image(self, i1, i2):
         assert i1.stock_image() == i2.stock_image()
-        assert i1.stock_image(category='space', width=128, height=128) == \
-            i2.stock_image(category='space', width=128, height=128)
 
     def test_image_by_keyword(self, i1, i2):
         assert i1.image_by_keyword() == i2.image_by_keyword()
         assert i1.image_by_keyword(keyword='food') == \
-            i2.image_by_keyword(keyword='food')
+               i2.image_by_keyword(keyword='food')
 
     def test_network_protocol(self, i1, i2):
         assert i1.network_protocol() == i2.network_protocol()
         assert i1.network_protocol(layer=Layer.PHYSICAL) == \
-            i2.network_protocol(layer=Layer.PHYSICAL)
+               i2.network_protocol(layer=Layer.PHYSICAL)
 
     def test_ip_v4(self, i1, i2):
         assert i1.ip_v4() == i2.ip_v4()
@@ -230,7 +230,7 @@ class TestSeededInternet(object):
     def test_content_type(self, i1, i2):
         assert i1.content_type() == i2.content_type()
         assert i1.content_type(mime_type=MimeType.APPLICATION) == \
-            i2.content_type(mime_type=MimeType.APPLICATION)
+               i2.content_type(mime_type=MimeType.APPLICATION)
 
     def test_http_status_code(self, i1, i2):
         assert i1.http_status_code() == i2.http_status_code()
@@ -241,12 +241,12 @@ class TestSeededInternet(object):
     def test_top_level_domain(self, i1, i2):
         assert i1.top_level_domain() == i2.top_level_domain()
         assert i1.top_level_domain(tld_type=TLDType.UTLD) == \
-            i2.top_level_domain(tld_type=TLDType.UTLD)
+               i2.top_level_domain(tld_type=TLDType.UTLD)
 
     def test_port(self, i1, i2):
         assert i1.port() == i2.port()
         assert i1.port(port_range=PortRange.REGISTERED) == \
-            i2.port(port_range=PortRange.REGISTERED)
+               i2.port(port_range=PortRange.REGISTERED)
 
     def test_category_of_website(self, i1, i2):
         assert i1.category_of_website() == i2.category_of_website()